force to return int instead  of long, else main ddoes not run

+++ src/main/java/application/model/Book.java
@@ -101,7 +101,7 @@ public class Book {
 
     // Getter und Setter für Werte
 
-    public int getBookId() {
+    public long getBookId() {
         return (int) bookId.get();
     }
